requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torch-pruning>=1.1.0
git+https://github.com/mit-han-lab/efficientvit.git
pandas>=2.0.0
Pillow>=9.0.0
matplotlib>=3.5.0
